+++ src/components/Courses/Layout/CourseLayoutMain.tsx
@@ -1,50 +1,14 @@
-import * as ReactDOM from "react-dom";
 import clsx from "clsx";
-import {
-  useCallback,
-  useState,
-  useRef,
-  useMemo,
-  ReactElement,
-  DetailedHTMLProps,
-  HTMLAttributes,
-  useEffect,
-  ReactNode,
-  MutableRefObject,
-} from "react";
+import { useCallback, useState, useRef, useMemo, useEffect } from "react";
 import { getMDXComponent } from "mdx-bundler/client";
-import ReactMarkdown from "react-markdown";
-import rehypeKatex from "rehype-katex";
-import remarkGfm from "remark-gfm";
-import remarkMath from "remark-math";
-import rehypeRaw from "rehype-raw";
 import TeX from "@matejmazur/react-katex";
-import {
-  AddressesType,
-  AnswerType,
-  CUSTOM_MATERIAL,
-  InputBoxElementType,
-  MatchBoxElementType,
-  MathFunction,
-  MathPoint,
-  OptionElementType,
-  StateType,
-} from "@/src/type";
-import {
-  checkChapterProgress,
-  evaluateMath,
-  getPracticeAnswer,
-  getPracticeId,
-  regexPracticeInput,
-} from "@/src/utils";
+import { AddressesType, AnswerType, StateType } from "@/src/type";
+import { checkChapterProgress } from "@/src/utils";
 import { Blockquote, Input, Loader, Paragraph } from "@/src/components";
 import { useRouter } from "next/router";
-import { BLOCKQUOTE_VARIANTS, COLORS } from "@/src/style";
-import { useToast } from "@/src/hooks";
-import { MatchBox } from "../Entity/Match";
+import { useCustom } from "@/src/hooks";
 
 import dynamic from "next/dynamic";
-import { Option } from "../Entity/Option/CourseEntityOption";
 
 const Graph = dynamic(() => import("../Entity/Graph/CourseEntityGraph"), {
   ssr: false,
@@ -73,7 +37,6 @@ export function CourseLayoutMain({
   stateAccept,
   stateLoading,
   stateSubmitted,
-  stateChecking,
   trueLoading,
   page,
   handleCheckAnswer,
@@ -85,45 +48,31 @@ export function CourseLayoutMain({
   const [accept, setAccept] = stateAccept;
   const [solved, setSolved] = stateSolved;
   const [submitted, setSubmmited] = stateSubmitted;
-  const answerInputBoxParentElement = useRef<InputBoxElementType[]>([]);
-  const matchParentElement = useRef<MatchBoxElementType[]>([]);
-  const optionParentElement = useRef<OptionElementType[]>([]);
-  const [active, setActive] = useState<any>(null);
-  const checking = stateChecking[0];
-
-  const matchRef = useRef<
-    Record<
-      string,
-      {
-        left: string;
-        right: string;
-        ready?: boolean;
-      }
-    >
-  >({});
+  const stateActive = useState<any>(null);
+  const active = stateActive[0];
   const inputRef = useRef<Record<string, boolean>>({});
-  const optionCount = useRef(0);
+
+  const {
+    handleOnePairMatch,
+    handleRenderAnswerBoxes,
+    handleRenderMatch,
+    handleRenderOptions,
+    handleConvertCodeToComponents,
+    handleRemoveAllCustomComponents,
+  } = useCustom({
+    handleCheckAnswer,
+    stateAccept,
+    stateActive,
+    stateAnswer,
+    stateLoading,
+    stateSolved,
+    stateSubmitted,
+    inputRef,
+  });
 
   const { practice } = addreses;
 
   const Content = useMemo(() => getMDXComponent(markdown), [markdown]);
-
-  const { addToast } = useToast();
-
-  const hanldeShowAnswerFeedback = useCallback(() => {
-    if (!checking || !submitted) return;
-
-    if (solved === 1) {
-      addToast({
-        phrase: "courseMaterialPracticeAnsweredCorrect",
-      });
-      return;
-    }
-
-    addToast({
-      phrase: "courseMaterialPracticeAnsweredIncorrect",
-    });
-  }, [addToast, checking, solved, submitted]);
 
   const userAnswerStatus = useCallback(
     (practiceId: string) => {
@@ -137,339 +86,9 @@ export function CourseLayoutMain({
     [answer, submitted, handleCheckAnswer]
   );
 
-  const renderCustomElement = useCallback(
-    (
-      parentElement: HTMLElement,
-      targetElement: ReactElement,
-      group: string,
-      id: string
-    ) => {
-      const vessel = document.createElement("div");
-      vessel.id = `${group}-${id}`;
-      vessel.classList.add(group);
-
-      parentElement.insertBefore(vessel, parentElement.nextSibling);
-
-      ReactDOM.render(targetElement, vessel);
-    },
-    []
-  );
-
-  const handleOnePairMatch = useCallback(() => {
-    if (active && active.type === "complete") {
-      setAnswer((prev) => ({
-        ...prev,
-        [active.id]: active.answer,
-      }));
-      setActive(null);
-    }
-  }, [active, setAnswer]);
-
-  const handleToggleOption = useCallback(
-    (questionId: string) => {
-      if (solved) return;
-
-      setAnswer((prev) => ({
-        ...prev,
-        [questionId]: prev[questionId] && prev[questionId] === "1" ? "0" : "1",
-      }));
-    },
-    [setAnswer, solved]
-  );
-
   useEffect(() => {
     handleOnePairMatch();
   }, [active, handleOnePairMatch]);
-
-  useEffect(() => {
-    console.log("KKKKKKKKKK");
-    console.log(answer);
-  }, [answer]);
-
-  const handleGetComponentForm = useCallback((str: string): ReactNode => {
-    if (str.match(/\$([^\$])*\$/g)) {
-      return <TeX>{str.slice(1, -1)}</TeX>;
-    }
-    return str;
-  }, []);
-
-  const handleSetActive = useCallback(
-    (type: "target" | "source", content: string, override: boolean = false) => {
-      setActive((prev: any) => {
-        if (!prev || override) {
-          return type === "target"
-            ? {
-                id: content,
-                type: "target",
-              }
-            : {
-                answer: content,
-                type: "source",
-              };
-        }
-        if (type === "target" && prev.type === "source") {
-          return {
-            id: content,
-            answer: prev.answer,
-            type: "complete",
-          };
-        } else if (type === "source" && prev.type === "target") {
-          return {
-            id: prev.id,
-            answer: content,
-            type: "complete",
-          };
-        } else {
-          return null;
-        }
-      });
-    },
-    []
-  );
-
-  const handleRemoveCustomComponents = useCallback((className: string) => {
-    const previousRenders = document.querySelectorAll(`.${className}`);
-    matchRef.current = {};
-    previousRenders.forEach((element) => {
-      /** @todos Something MIGHT be wrong with this unmount method. */
-      element.parentElement?.removeChild(element);
-    });
-  }, []);
-
-  const handleRemoveAllCustomComponents = useCallback(() => {
-    // Object.entries(customElementQueue.current).forEach(([key, _]) => {
-    //   JXG.JSXGraph.freeBoard(key);
-    // });
-    Object.values(CUSTOM_MATERIAL).forEach((group) => {
-      handleRemoveCustomComponents(group);
-    });
-
-    inputRef.current = {};
-  }, [handleRemoveCustomComponents]);
-
-  const handleRemoveAnswer = useCallback(
-    (key: string) => {
-      setAnswer((prev) => ({
-        ...prev,
-        [key]: undefined,
-      }));
-    },
-    [setAnswer]
-  );
-
-  const handleCheckIfAlreadyMatched = useCallback(
-    (
-      currentKey: any,
-      currentAnswer: any,
-      onKeyMatch: null | ((key: string) => void),
-      onValueMatch: null | ((key: string) => void)
-    ) => {
-      if (solved === 1) return;
-      Object.entries(answer).forEach(([key, value]) => {
-        if (onKeyMatch && key === currentKey) {
-          onKeyMatch(key);
-        } else if (onValueMatch && value === currentAnswer) {
-          onValueMatch(key);
-        }
-      });
-    },
-    [answer, solved]
-  );
-
-  const handleClickMatchedCard = useCallback(
-    (practiceId: string) =>
-      handleCheckIfAlreadyMatched(practiceId, null, handleRemoveAnswer, null),
-    [handleCheckIfAlreadyMatched, handleRemoveAnswer]
-  );
-
-  const handleClickDrop = useCallback(
-    (practiceId: string) => {
-      let terminate = false;
-
-      handleCheckIfAlreadyMatched(
-        practiceId,
-        null,
-        (key: string) => {
-          handleRemoveAnswer(key);
-          terminate = true;
-        },
-        null
-      );
-
-      if (active && active.type === "target") {
-        if (active.id === practiceId) {
-          setActive(null);
-        } else {
-          handleSetActive("target", practiceId, true);
-        }
-        terminate = true;
-      }
-
-      if (terminate) return;
-
-      handleSetActive("target", practiceId);
-    },
-    [active, handleCheckIfAlreadyMatched, handleRemoveAnswer, handleSetActive]
-  );
-
-  const handleClickUnmatchedCard = useCallback(
-    (right: string) => {
-      let terminate = false;
-
-      handleCheckIfAlreadyMatched(null, right, null, (key: string) => {
-        handleRemoveAnswer(key);
-
-        if (active && active.id) setActive(null);
-        terminate = true;
-      });
-
-      if (active && active.type === "source") {
-        if (active.answer === right) {
-          setActive(null);
-        } else {
-          handleSetActive("source", right, true);
-        }
-        terminate = true;
-      }
-      if (terminate) return;
-
-      handleSetActive("source", right);
-    },
-    [active, handleCheckIfAlreadyMatched, handleRemoveAnswer, handleSetActive]
-  );
-
-  const renderOption = useCallback(
-    (practiceId: string, content: string) => {
-      const identifier = `Option-${practiceId}`;
-
-      return (
-        <Option
-          id={identifier}
-          content={content}
-          selected={Boolean(answer[practiceId] && answer[practiceId] === "1")}
-          onSelect={() => {
-            handleToggleOption(practiceId);
-          }}
-          solved={solved}
-        />
-      );
-    },
-    [answer, handleToggleOption, solved]
-  );
-
-  const renderMatchBox = useCallback(
-    (practiceId: string, left: string, right: string) => {
-      const identifier = `MatchBox-${practiceId}`;
-
-      //.replaceAll("\\{", "{").replaceAll("\\}", "}")
-      if (!matchRef.current[practiceId])
-        matchRef.current[practiceId] = {
-          left: left.replaceAll("\\{", "{").replaceAll("\\}", "}"),
-          right: right.replaceAll("\\{", "{").replaceAll("\\}", "}"),
-          ready: true,
-        };
-
-      return (
-        <MatchBox
-          // key={identifier}
-          id={identifier}
-          practiceId={practiceId}
-          active={active}
-          answer={answer}
-          solved={solved}
-          left={matchRef.current[practiceId].left}
-          right={matchRef.current[practiceId].right}
-          handleClickMatchedCard={() => handleClickMatchedCard(practiceId)}
-          handleClickDrop={() => handleClickDrop(practiceId)}
-          handleClickUnmatchedCard={() => handleClickUnmatchedCard(right)}
-          handleGetComponentForm={handleGetComponentForm}
-        />
-      );
-    },
-    [
-      active,
-      answer,
-      handleClickDrop,
-      handleClickMatchedCard,
-      handleClickUnmatchedCard,
-      handleGetComponentForm,
-      solved,
-    ]
-  );
-
-  const renderAnswerBox = useCallback(
-    (practiceId: string) => (
-      <Input
-        key={`InputBox-${practiceId}`}
-        id={`InputBox-${practiceId}`}
-        onBlur={(e) => {
-          if (answer !== accept) {
-            setSubmmited(false);
-            setAnswer((prev) => ({
-              ...prev,
-              [practiceId]: e.target.value,
-            }));
-          }
-        }}
-        defaultValue={answer[practiceId]}
-        disabled={solved === 1 || userAnswerStatus(practiceId) === "success"}
-        state={submitted ? userAnswerStatus(practiceId) : undefined}
-      />
-    ),
-    [
-      answer,
-      solved,
-      userAnswerStatus,
-      submitted,
-      accept,
-      setSubmmited,
-      setAnswer,
-    ]
-  );
-
-  const handleRenderAnswerBoxes = useCallback(() => {
-    handleRemoveCustomComponents("InputBox");
-
-    answerInputBoxParentElement.current.forEach(({ parentElement, string }) => {
-      const currentId = getPracticeId(string);
-      if (currentId) {
-        renderCustomElement(
-          parentElement,
-          renderAnswerBox(currentId),
-          CUSTOM_MATERIAL["input"],
-          currentId
-        );
-      }
-    });
-  }, [handleRemoveCustomComponents, renderAnswerBox, renderCustomElement]);
-
-  const handleRenderMatch = useCallback(() => {
-    handleRemoveCustomComponents("MatchBox");
-
-    matchParentElement.current.forEach(({ parentElement, pair, id }) => {
-      const left = pair[0];
-      const right = pair[1];
-      renderCustomElement(
-        parentElement,
-        renderMatchBox(id, left, right),
-        CUSTOM_MATERIAL["match"],
-        id
-      );
-    });
-  }, [handleRemoveCustomComponents, renderCustomElement, renderMatchBox]);
-
-  const handleRenderOptions = useCallback(() => {
-    handleRemoveCustomComponents("Option");
-
-    optionParentElement.current.forEach(({ parentElement, content, id }) => {
-      renderCustomElement(
-        parentElement,
-        renderOption(id, content),
-        CUSTOM_MATERIAL["option"],
-        id
-      );
-    });
-  }, [handleRemoveCustomComponents, renderCustomElement, renderOption]);
 
   useEffect(() => {
     handleRenderAnswerBoxes();
@@ -548,172 +167,6 @@ export function CourseLayoutMain({
     handleGetExistingAnswerIfAny();
   }, [accept, handleGetExistingAnswerIfAny]);
 
-  const handleConvertCodeToComponents = useCallback(() => {
-    if (!loading) return;
-
-    const container = document.getElementsByClassName(
-      "CourseMaterial_wrapper"
-    )[0];
-
-    handleRemoveAllCustomComponents();
-
-    const elements = document.querySelectorAll(
-      ".CourseMaterial_content .CustomMaterialInvoker"
-    );
-
-    let inputElementsRendered = 0;
-
-    let answerKeys = {};
-
-    answerInputBoxParentElement.current = [];
-    matchParentElement.current = [];
-    optionParentElement.current = [];
-
-    elements.forEach((element, index) => {
-      const string = element.innerHTML;
-      const parentElement = element.parentElement;
-
-      if (container && parentElement && string.match(/\<Practice/g)) {
-        const detectedPair = string.toString().split("@");
-
-        if (detectedPair && detectedPair.length === 4) {
-          const [tag, id, left, right] = detectedPair;
-          matchParentElement.current = [
-            ...matchParentElement.current,
-            {
-              parentElement,
-              pair: [left, right],
-              id,
-            },
-          ];
-
-          answerKeys = {
-            ...answerKeys,
-            [id]: `${right}`,
-          };
-
-          inputElementsRendered++;
-        }
-      }
-
-      if (container && parentElement && string.match(/\[match\]/g)) {
-        const detectedPair = string.toString().split("@");
-
-        if (detectedPair && detectedPair.length === 4) {
-          const [tag, id, left, right] = detectedPair;
-          matchParentElement.current = [
-            ...matchParentElement.current,
-            {
-              parentElement,
-              pair: [left, right],
-              id,
-            },
-          ];
-
-          answerKeys = {
-            ...answerKeys,
-            [id]: `${right}`,
-          };
-
-          inputElementsRendered++;
-        }
-      }
-
-      if (container && parentElement && string.match(/\[option\]/g)) {
-        const detectedPair = string.toString().split("@");
-
-        if (detectedPair && detectedPair.length === 4) {
-          const [tag, id, content, truth] = detectedPair;
-          optionParentElement.current = [
-            ...optionParentElement.current,
-            {
-              parentElement,
-              content,
-              id,
-              truth: Number(truth),
-            },
-          ];
-
-          answerKeys = {
-            ...answerKeys,
-            [id]: `${truth}`,
-          };
-
-          inputElementsRendered++;
-        }
-      }
-    });
-
-    // Object.entries(customElementQueue.current).forEach(
-    //   ([key, { containerId, element, props }]) => {
-    //     const target = document.getElementById(containerId);
-    //     if (target)
-    //       renderCustomElement(
-    //         target,
-    //         <div className="GraphContainer">
-    //           <Graph id={key} {...props} />
-    //         </div>,
-    //         CUSTOM_MATERIAL["graph"],
-    //         key
-    //       );
-    //   }
-    // );
-
-    if (matchParentElement.current.length > 0 && solved !== 1) {
-      matchParentElement.current.sort(() => Math.random() - 0.5);
-      let randomizedPairs = matchParentElement.current.map(({ pair }) => {
-        return pair[1];
-      });
-
-      randomizedPairs.sort(() => Math.random() - 0.5);
-      matchParentElement.current = matchParentElement.current.map(
-        (parent, idx: number) => ({
-          ...parent,
-          pair: [parent.pair[0], randomizedPairs[idx]],
-        })
-      );
-    }
-
-    if (optionParentElement.current.length > 0 && solved !== 1) {
-      if (!solved) optionParentElement.current.sort(() => Math.random() - 0.5);
-
-      const defaultTruth = optionParentElement.current
-        .map((option) => option.id)
-        .reduce(
-          (prev, curr) => ({
-            ...prev,
-            [curr]: "0",
-          }),
-          {}
-        );
-      setAnswer((prev) => ({
-        ...prev,
-        ...defaultTruth,
-      }));
-    }
-
-    if (inputElementsRendered > 0 && Object.values(answerKeys).length > 0) {
-      setSolved(0);
-      setAccept((prev) => ({
-        ...prev,
-        ...answerKeys,
-      }));
-      handleRenderAnswerBoxes();
-      handleRenderMatch();
-      handleRenderOptions();
-    }
-  }, [
-    loading,
-    handleRemoveAllCustomComponents,
-    solved,
-    setAnswer,
-    setSolved,
-    setAccept,
-    handleRenderAnswerBoxes,
-    handleRenderMatch,
-    handleRenderOptions,
-  ]);
-
   const handlePrepareNewPage = useCallback(() => {
     if (loading) {
       handleRemoveAllCustomComponents();
@@ -730,85 +183,6 @@ export function CourseLayoutMain({
   useEffect(() => {
     handlePrepareNewPage();
   }, [page, handlePrepareNewPage]);
-
-  const handleCheckForCodeInvokedElements = useCallback(
-    (element: any, specific: string = "") => {
-      if (specific !== "") {
-        return element.toString().match(new RegExp(`\[${specific}\]`));
-      }
-      return element.toString().match(/\[[^\]]*\]/g);
-    },
-    []
-  );
-
-  const handlePreTransformCode = useCallback(
-    ({ node, children, ...props }: any) => {
-      return handleCheckForCodeInvokedElements(children) ? (
-        <span className="CustomMaterialInvoker hidden">{children}</span>
-      ) : (
-        <code>{children}</code>
-      );
-    },
-    [handleCheckForCodeInvokedElements]
-  );
-
-  const handleGetBlockquoteWithoutTags = useCallback((element: any) => {
-    const regex = /\#([^\#])*\#/g;
-
-    return element.map((str: any) => {
-      if (typeof str === "object") {
-        const newChildren = str.props.children.map((x: any) =>
-          typeof x === "string" ? x.replace(regex, "") : x
-        );
-        return {
-          ...str,
-          props: {
-            ...str.props,
-            children: newChildren,
-          },
-        };
-      }
-      return str.replace(regex, "");
-    });
-  }, []);
-
-  const handleCheckForSpecialBlockquote = useCallback(
-    (element: any, type: string) => {
-      const regex = new RegExp(`\#${type}\#`);
-
-      return element.some((str: any) => {
-        if (typeof str === "object") {
-          return str.props.children.some((x: any) =>
-            typeof x === "string" ? x.match(regex) : false
-          );
-        }
-        return str.match(regex);
-      });
-    },
-    []
-  );
-
-  const handlePreTransformBlockquote = useCallback(
-    ({ node, children, ...props }: any) => {
-      const taglessChildren = handleGetBlockquoteWithoutTags(children);
-
-      let quoteProps = {};
-
-      const variants = BLOCKQUOTE_VARIANTS.filter((c) =>
-        handleCheckForSpecialBlockquote(children, c)
-      );
-
-      if (variants.length === 1) {
-        quoteProps = {
-          className: variants[0] === "explanation" && !solved && "hidden",
-          variant: variants[0],
-        };
-      }
-
-      return <Blockquote {...quoteProps}>{taglessChildren}</Blockquote>;
-    },
-    [handleCheckForSpecialBlockquote, handleGetBlockquoteWithoutTags, solved]
-  );
 
   const handleRouteChangeStart = useCallback(() => {
     onChapterChange && onChapterChange();
@@ -919,15 +293,6 @@ export function CourseLayoutMain({
                 <span className="CustomMaterialInvoker hidden">{`[option]@${id}@${content}@${
                   truth ? 1 : 0
                 }`}</span>
-                // <Option
-                //   selected={
-                //     !answer[id] && answer[id] === "t"
-                //   }
-                //   onSelect={() => {
-                //     handleToggleOption(id);
-                //   }}
-                // 	content={content}
-                // />
               ),
             }}
           />
